requirements.txt
# This file contains the Python dependencies for the IRCTC Tatkal Bot.
# To install them, run the following command in your terminal:
# pip install -r requirements.txt

# --- Core Application ---
streamlit
selenium
requests

# --- OCR Dependencies ---
easyocr
torch
opencv-python-headless
numpy
Pillow
